Rename ProductSwitcherMenu with SwitchProductMenu
