--- webapp/channels/src/components/global_header/left_controls/product_switcher_menu/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-export {ProductSwitcherMenu as default} from './product_switcher_menu';
-
-export {ELEMENT_ID_FOR_PRODUCT_SWITCHER_MENU, ELEMENT_ID_FOR_PRODUCT_SWITCHER_MENU_BUTTON} from './product_switcher_menu';
