+++ webapp/channels/src/components/global_header/left_controls/switch_product_menu/index.ts
@@ -1,0 +1,8 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+export {
+    SwitchProductMenu as default,
+    ELEMENT_ID_FOR_SWITCH_PRODUCT_MENU,
+    ELEMENT_ID_FOR_SWITCH_PRODUCT_MENU_BUTTON,
+} from './switch_product_menu';
