+++ webapp/channels/src/components/global_header/left_controls/switch_product_menu/switch_product_cloud_limits_footer.tsx
