+++ webapp/channels/src/components/global_header/left_controls/switch_product_menu/switch_product_system_console_menuitem.tsx
