+++ webapp/channels/src/components/onboarding_tasks/onboarding_tasks_manager.tsx
@@ -26,7 +26,7 @@ import Handshake from 'components/common/svg_images_components/handshake_svg';
 import Phone from 'components/common/svg_images_components/phone_svg';
 import Security from 'components/common/svg_images_components/security_svg';
 import Sunglasses from 'components/common/svg_images_components/sunglasses_svg';
-import {ELEMENT_ID_FOR_PRODUCT_SWITCHER_MENU_BUTTON} from 'components/global_header/left_controls/product_switcher_menu';
+import {ELEMENT_ID_FOR_SWITCH_PRODUCT_MENU_BUTTON} from 'components/global_header/left_controls/switch_product_menu';
 import LearnMoreTrialModal from 'components/learn_more_trial_modal/learn_more_trial_modal';
 import {openMenu} from 'components/menu';
 import {
@@ -239,7 +239,7 @@ export const useHandleOnBoardingTaskTrigger = () => {
             break;
         }
         case OnboardingTasksName.VISIT_SYSTEM_CONSOLE: {
-            openMenu(ELEMENT_ID_FOR_PRODUCT_SWITCHER_MENU_BUTTON);
+            openMenu(ELEMENT_ID_FOR_SWITCH_PRODUCT_MENU_BUTTON);
             dispatch(setShowOnboardingVisitConsoleTour(true));
             handleSaveData(taskName, TaskNameMapToSteps[taskName].STARTED);
             break;
